requirements.txt
pylint
PyJWT
pytest
flask==1.1.4
Flask-SQLAlchemy
Flask-Sockets
websocket-client # just for ut
